Add changelog entries and remove FIXME
We've determined the FIXME isn't very practical, since Jinja2 doesn't maintain a distinction between globals/vars/locals in
a consistent manner when new contexts are involved. While it might be possible to keep them separate with enough work, doing so
would be fragile and subject to breakage when Jinja2 changed. It would also be extremely difficult to understand and verify
correctness.

It's much easier to make all globals/locals/vars always templated.